pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paintbynumbers"
version = "1.0.0"
description = "Paint by Numbers Generator - Convert images to paint-by-numbers with numbered regions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Paint by Numbers Generator Contributors"}
]
keywords = ["paint-by-numbers", "image-processing", "svg", "k-means", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "Pillow>=10.0.0,<11.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "svgwrite>=1.4.0,<2.0.0",
    "click>=8.1.0,<9.0.0",
    "cairosvg>=2.7.0,<3.0.0",
    "tqdm>=4.65.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "types-Pillow",
    "types-tqdm",
]

[project.scripts]
paintbynumbers = "paintbynumbers.cli.main:main"
paintbynumbers-config = "paintbynumbers.cli.main:init_config"

[project.urls]
Homepage = "https://github.com/Nuopel/paintbynumbersgenerator"
Repository = "https://github.com/Nuopel/paintbynumbersgenerator"
Issues = "https://github.com/Nuopel/paintbynumbersgenerator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
paintbynumbers = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=paintbynumbers",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["src/paintbynumbers"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "svgwrite.*",
    "cairosvg.*",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["paintbynumbers"]
